feat: add styles for FilterBar and add useCallback
