+++ src/components/FilterBar/FilterBar.module.css
@@ -1,0 +1,17 @@
+.filterBar {
+    display: flex;
+    gap: 10px;
+    margin-bottom: 20px;
+}
+
+.filterBar select,
+.filterBar input {
+    padding: 5px;
+    font-size: 1em;
+}
+
+.filterBar button {
+    padding: 5px 10px;
+    font-size: 1em;
+    cursor: pointer;
+}
